refactor(pi07): rename low_level_planner -> low_level (it's not a planner)
The low-level component of π0.7 is a flow-matching action generator
with FAST discrete action prediction, not a planner. The high-level
component IS a planner (predicts subtask strings autoregressively) and
keeps its naming. This PR aligns the source-side naming with that
distinction. Closes #247.

Changes:

- Move src/opentau/policies/pi07/low_level_planner/ -> low_level/.
- Rename classes:
  - PI07LowLevelPlannerConfig    -> PI07LowLevelConfig
  - PI07LowLevelPlannerPolicy    -> PI07LowLevelPolicy
  - PI07LowLevelPlannerFlowMatching -> PI07LowLevelFlowMatching
- Update factory.py, policies/__init__.py, gemma3_with_expert.py,
  high_level_planner docstring, and the test suite (test_pi07_cpu.py,
  test_pi07_attention_layout.py, test_pi07_low_level.py,
  test_pi07_video_encoder_cpu.py, test_available.py).
- The TestPI07LowLevelPlannerAttentionLayout class in
  test_pi07_attention_layout.py drops "Planner" too.

Backward compatibility:

A deprecation shim is left at the old low_level_planner/ path that
re-exports the new classes under the old names with a
DeprecationWarning at package-import time. This keeps existing user
code, pickled checkpoints (config classes), and any draccus configs
that pinned the old import path working through one release cycle.

The shim is exercised by tests/policies/test_pi07_low_level_shim.py:
verifies the warning fires, and that PI07LowLevelPlanner* names are
the same class objects as the new PI07LowLevel* names.

The policy-type strings in factory.py (`pi07_low_level`,
`pi07_high_level`) are unchanged, so draccus `policy.type` configs
keep working without edits.

Out of scope:

- The legacy pi07_paligemma/low_level_planner/ package and its
  PI07lowlevelPlannerConfig / PI07LowLevelPlannerPolicy classes are
  intentionally untouched. That is a separate retired package.
- High-level naming is preserved everywhere.

Verification:

- pre-commit run --all-files: clean
- pytest -m "not gpu" -n auto tests/policies/test_pi07_attention_layout.py
  tests/policies/test_pi07_cpu.py tests/policies/test_pi07_video_encoder_cpu.py
  tests/policies/test_pi07_low_level_shim.py tests/test_available.py:
  62 passed in 54s.
- Full CPU suite (matching cpu_test.yml): 903 passed, 13 skipped, plus
  2 pre-existing timing flakes in tests/utils/test_utils_benchmark.py
  (timing tolerance on a 50ms sleep — unrelated to this PR).
