+++ src/opentau/policies/__init__.py
@@ -26,7 +26,7 @@ from .pi05_mem.configuration_pi05 import PI05MemConfig as PI05MemConfig
 from .pi07.high_level_planner.configuration_pi07_high_level import (
     PI07HighLevelPlannerConfig as PI07HighLevelPlannerConfig,
 )
-from .pi07.low_level_planner.configuration_pi07_low_level import (
-    PI07LowLevelPlannerConfig as PI07LowLevelPlannerConfig,
+from .pi07.low_level.configuration_pi07_low_level import (
+    PI07LowLevelConfig as PI07LowLevelConfig,
 )
 from .value.configuration_value import ValueConfig as ValueConfig
